% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_clustering.R
\name{select_r}
\alias{select_r}
\title{Number of clusters selection}
\usage{
select_r(Y, r_range)
}
\arguments{
\item{Y}{array, order-3 Gaussian tensor observation}

\item{r_range}{vector, candidates for the number of clusters}
}
\value{
a list containing the following:

\code{r} {number, the number of clusters among the candidates with minimal BIC value}

\code{bic} {vector, the BIC value for each candidiate}
}
\description{
Estimate the number of clusters in the degree-corrected tensor block model based on BIC criterion. The choice of BIC
aims to balance between the goodness-of-fit for the data and the degree of freedom in the population model.
This function is restricted to the symmetric case with order-3 Gaussian tensor observation.
}
\examples{

#' seed = 1
p = 30
r = 3
delta = 0.5
s_min = 0.05
s_max = 1
dist = "normal"
theta_dist = "pareto"
alpha = 4
beta = 3/4
sigma = 0.2

data = sim_hDCBM_network(seed = seed, p, r,  delta = delta,
 s_min = s_min, s_max = s_max, dist =  dist, sigma = sigma,
 theta_dist = theta_dist, alpha = alpha, beta = beta, imat = F)

r_range = 2:5

selection = select_r(data$Y, r_range)
}
