% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tensor_clustering.R
\name{wkmeans}
\alias{wkmeans}
\title{Weighted higher-order initialization}
\usage{
wkmeans(Y, r, asymm = F)
}
\arguments{
\item{Y}{order-3 array or matrix, order-3 tensor or matrix observation}

\item{r}{vector, the number of clusters on each model; see "details"}

\item{asymm}{logic variable, if "TRUE", assume the clustering assignment differs in different modes; if "FALSE", assume all the modes share the same clustering assignment}
}
\value{
a list containing the following:

\code{z0} { vector ( if \code{asymm = F} ) or a list of vectors ( if \code{asymm = T} ) recording the estimated clustering assignment }

\code{s0} { vector ( if \code{asymm = F} ) or a list of vectors ( if \code{asymm = T} ) recording the index of degenerate entities with random clustering assignment}
}
\description{
Weighted higher-order initialization for multiway spherical clustering under degree-corrected tensor block model.
This function takes the tensor/matrix observation, the number of clusters, and a logic variable indicating the symmetry
of the tensor as input. The output is the estimated clustering assignment.
}
\details{
For the number of clusters \code{r}, the length of \code{r} is no longer than 3 with integers larger than 1;
the input \code{r = r0} is equal to input \code{r = rep(r0,3)} for any integer \code{r0 > 1};
matrix observation only accounts the frist two elements and requires all the elements in \code{r} are identicial.
}
\examples{
seed = 1
p = 30
r = 3
delta = 0.5
s_min = 0.05
s_max = 1
dist = "normal"
theta_dist = "pareto"
alpha = 4
beta = 3/4
sigma = 0.2

data = sim_hDCBM_network(seed = seed, p, r,  delta = delta,
 s_min = s_min, s_max = s_max, dist =  dist, sigma = sigma,
 theta_dist = theta_dist, alpha = alpha, beta = beta, imat = F)

initialization = wkmeans(data$Y,r = 3, asymm = F)
}
